refactor(targets): extract useTargetsPageFilters hook and target-search-helpers from TargetsPage (CU-32)
629L → ~250L TargetsPage; filter/search/sort/reveal state in hook; normalizeDesig+matchesSearch in dedicated helpers file. Skip useRevealChunking: virtualization in useTargetsTableRows already absorbed it.
